Replaced the time measurement funcion from SDL_GetTicks() to sys/time.h to measure time in Microseconds

+++ src/lib.h
@@ -20,13 +20,13 @@ char* game_signature;           // Game signature to apply quirks
 // -------------------------------------- Functions ------------------------------------- //
 
 // Ticker - Use with actions that should be executed each second = 1hz
-bool ticker_second(unsigned int lastTime, unsigned int currentTime);
+bool ticker_second(long lastTime, long currentTime);
 
 // Ticker - Use with actions that should be executed 60 times per second = 60hz
-bool ticker_fps(unsigned int lastTime, unsigned int currentTime);
+bool ticker_fps(long lastTime, long currentTime);
 
 // Ticker - se with actions that should be executed 500 times per second = 500hz
-bool ticker_cpu(unsigned int lastTime, unsigned int currentTime);
+bool ticker_cpu(long lastTime, long currentTime);
 
 // Load rom into memory
 void load_rom(char* filename, unsigned char *mem, unsigned int mem_size);
